+++ code/Sitecore.TestRunnerJS/Install.ps1
@@ -2,7 +2,9 @@ param($installPath, $toolsPath, $package, $project)
 
 $projectPath = Split-Path $project.FileName
 
-Write-Host "TestRunnerJS files location: $toolsPath"
-Write-Host "Installation location: $projectPath"
+if (Test-Path "$projectPath\App_Config") {
+	Write-Host "TestRunnerJS files location: $toolsPath"
+	Write-Host "Installation location: $projectPath"
 
-Copy-Item -Path "$toolsPath\*" -Destination "$projectPath" -Exclude "*.ps1" -recurse -Force
+	Copy-Item -Path "$toolsPath\*" -Destination "$projectPath" -Exclude "*.ps1" -recurse -Force
+}
